Complete writing related functions
We have fully implemented write functions that were not previously fully implemented.
We added the ability to check whether the new file was created, deleted the existing file, or modified the existing file by examining the change of the fat area in byte units.
In addition, the union-find data structure was used to store the start cluster of each file in a linear time in any cluster in the same file.