Delete the base code
The existing fat32 file system code supported a variety of powerful features,
but there were too many features not needed to build a virtual file system.
In addition, since it was difficult to debug with a large amount of code,
we had to delete the underlying code and re-implement only the necessary functions.