Support case sensitive function
The reason for the fat32 directory entry is unknown, but it is not case-sensitive, only capital letters. But we do not need to do that because we use a virtual file system. Therefore, the file name is written in the entry as it is, so that the filename can be case-sensitive.